pyproject.toml
[tool.isort]
profile = "black"
src_paths = ["goath2", "tests"]

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_defs = true
explicit_package_bases = true
files = ["**/*.py"]
namespace_packages = true
show_error_codes = true

[tool.poetry]
name = "goath2"
version = "0.1.0-alpha"
description = ""
authors = ["Nikos Leivadaris <nikosleiv@gmail.com>"]
license = "MIT"
homepage = "https://github.com/nikosl/goath2"
repository = "https://github.com/nikosl/goath2"
keywords = ["google"]
readme = "README.md"
include = ["goath2/py.typed"]

[tool.poetry.scripts]
goath2 = "goath2.console:main"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.4"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"

[tool.poetry.dev-dependencies]
black = {version = "*", allow-prereleases = true}
flake8 = "*"
isort = "^5.6"
mkdocs-material = "^8"
mypy = ">0.900,<1"
pre-commit = "^2.8"
pytest = "^6.2"
pytest-cov = "^3"
pytest-mock = "^3.3"
python-semantic-release = "^7.27.0"

[tool.pytest.ini_options]
addopts = "-q"
minversion = "6.0"
testpaths = ["tests"]

[tool.semantic_release]
version_variable = [
    "pyproject.toml:tool.poetry.version"
]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "poetry build"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
